Adjusted the code that will suit final test

+++ gyros/FXAS21002C_Arduino_Library-master/FXAS21002C.cpp
@@ -133,7 +133,7 @@ float FXAS21002C::sensitivity(void)
 }
 
 // returns the current output data rate as a int (Hz)
-int FXAS21002C::getODR(void)
+double FXAS21002C::getODR(void)
 {
 	switch (gyroODR)
 	{
